+++ ecmascript-hot/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'gadicc:ecmascript-hot',
   // version: '0.4.1-rc.10'    // core version, KEEP UPDATED
-  version: '0.0.12-rc.8',
+  version: '0.0.13-rc.10',
   summary: 'Replacement ecmascript package providing react hotloading',
   git: 'https://github.com/gadicc/meteor-react-hotloader',
   documentation: '../README.md'
@@ -14,9 +14,9 @@ Package.registerBuildPlugin({
 });
 
 Package.onUse(function (api) {
-  api.versionsFrom('1.3-rc.3');
+  api.versionsFrom('1.3-rc.4');
   api.use('isobuild:compiler-plugin@1.0.0');
-  api.use('gadicc:babel-compiler-hot@0.0.13-rc.8');
+  api.use('gadicc:babel-compiler-hot@0.0.15-rc.10');
 
   api.imply('modules');
   api.imply('ecmascript-runtime');